+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added python package for post-processing ObsFcstAna output.
+
 ### Changed
 
 - Specify only ntasks_model for SLURM resource request.